Use injection to build Settings' PermissionsListener.
