+++ app/src/main/java/com/itsronald/twenty2020/settings/injection/SettingsComponent.kt
@@ -3,6 +3,7 @@ package com.itsronald.twenty2020.settings.injection
 import com.itsronald.twenty2020.base.Activity
 import com.itsronald.twenty2020.data.ResourceComponent
 import com.itsronald.twenty2020.settings.SettingsActivity
+import com.karumi.dexter.listener.single.PermissionListener
 import dagger.Component
 
 @Activity
@@ -11,5 +12,6 @@ import dagger.Component
         modules = arrayOf(SettingsModule::class)
 )
 interface SettingsComponent {
+    fun permissionsListener(): PermissionListener
     fun inject(settingsActivity: SettingsActivity)
 }
